feat: - se implementa el envio de snapshots: se envian y reciben snapshots que incluyen caracteristicas del juego (map_dimensions, supplies, disparos, players).
refactor:
- constantes: game, common, physics.

se hace pre-commit

+++ common_src/snapshots/snapshot_supply.h
@@ -6,15 +6,15 @@
 class SupplySnapshot {
 public:
     SupplyType type;
-    float pos_x;
-    float pos_y;
+    int pos_x;
+    int pos_y;
     char id;
     SupplyState state;
-    float width;
-    float height;
+    int width;
+    int height;
 
-    SupplySnapshot(SupplyType type, float pos_x, float pos_y, char id, SupplyState state,
-                   float width, float height):
+    SupplySnapshot(SupplyType type, int pos_x, int pos_y, char id, SupplyState state, int width,
+                   int height):
             type(type),
             pos_x(pos_x),
             pos_y(pos_y),
